refactored matching code, rewrote match_intervals. implements #706
