+++ librosa/util/__init__.py
@@ -66,6 +66,7 @@ File operations
 
 from .utils import *  # pylint: disable=wildcard-import
 from .files import *  # pylint: disable=wildcard-import
+from .matching import *  # pylint: disable=wildcard-import
 from .deprecation import *  # pylint: disable=wildcard-import
 from . import decorators
 from . import exceptions
