+++ librosa/util/utils.py
@@ -12,8 +12,6 @@ from numpy.lib.stride_tricks import as_strided
 from .. import cache
 from .exceptions import ParameterError
 
-from numba import jit
-
 # Constrain STFT block sizes to 256 KB
 MAX_MEM_BLOCK = 2**8 * 2**10
 
@@ -22,7 +20,6 @@ __all__ = ['MAX_MEM_BLOCK',
            'valid_audio', 'valid_int', 'valid_intervals',
            'fix_frames',
            'axis_sort', 'localmax', 'normalize',
-           'match_intervals', 'match_events',
            'peak_pick',
            'sparsify_rows',
            'roll_sparse',
@@ -777,226 +774,6 @@ def normalize(S, norm=np.inf, axis=0, threshold=None, fill=None):
         Snorm[:] = S / length
 
     return Snorm
-
-
-def match_intervals(intervals_from, intervals_to):
-    '''Match one set of time intervals to another.
-
-    This can be useful for tasks such as mapping beat timings
-    to segments.
-
-    .. note:: A target interval may be matched to multiple source
-      intervals.
-
-    Parameters
-    ----------
-    intervals_from : np.ndarray [shape=(n, 2)]
-        The time range for source intervals.
-        The `i` th interval spans time `intervals_from[i, 0]`
-        to `intervals_from[i, 1]`.
-        `intervals_from[0, 0]` should be 0, `intervals_from[-1, 1]`
-        should be the track duration.
-
-    intervals_to : np.ndarray [shape=(m, 2)]
-        Analogous to `intervals_from`.
-
-    Returns
-    -------
-    interval_mapping : np.ndarray [shape=(n,)]
-        For each interval in `intervals_from`, the
-        corresponding interval in `intervals_to`.
-
-    See Also
-    --------
-    match_events
-
-    Raises
-    ------
-    ParameterError
-        If either array of input intervals is not the correct shape
-    '''
-
-    if len(intervals_from) == 0 or len(intervals_to) == 0:
-        raise ParameterError('Attempting to match empty interval list')
-
-    # Verify that the input intervals has correct shape and size
-    valid_intervals(intervals_from)
-    valid_intervals(intervals_to)
-
-    # The overlap score of a beat with a segment is defined as
-    #   max(0, min(beat_end, segment_end) - max(beat_start, segment_start))
-    output = np.empty(len(intervals_from), dtype=np.int)
-
-    n_rows = int(MAX_MEM_BLOCK / (len(intervals_to) * intervals_to.itemsize))
-    n_rows = max(1, n_rows)
-
-    for bl_s in range(0, len(intervals_from), n_rows):
-        bl_t = min(bl_s + n_rows, len(intervals_from))
-        tmp_from = intervals_from[bl_s:bl_t]
-
-        starts = np.maximum.outer(tmp_from[:, 0], intervals_to[:, 0])
-        ends = np.minimum.outer(tmp_from[:, 1], intervals_to[:, 1])
-        score = np.maximum(0, ends - starts)
-
-        output[bl_s:bl_t] = np.argmax(score, axis=-1)
-
-    return output
-
-
-def match_events(events_from, events_to, left=True, right=True):
-    '''Match one set of events to another.
-
-    This is useful for tasks such as matching beats to the nearest
-    detected onsets, or frame-aligned events to the nearest zero-crossing.
-
-    .. note:: A target event may be matched to multiple source events.
-
-    Examples
-    --------
-    >>> # Sources are multiples of 7
-    >>> s_from = np.arange(0, 100, 7)
-    >>> s_from
-    array([ 0,  7, 14, 21, 28, 35, 42, 49, 56, 63, 70, 77, 84, 91,
-           98])
-    >>> # Targets are multiples of 10
-    >>> s_to = np.arange(0, 100, 10)
-    >>> s_to
-    array([ 0, 10, 20, 30, 40, 50, 60, 70, 80, 90])
-    >>> # Find the matching
-    >>> idx = librosa.util.match_events(s_from, s_to)
-    >>> idx
-    array([0, 1, 1, 2, 3, 3, 4, 5, 6, 6, 7, 8, 8, 9, 9])
-    >>> # Print each source value to its matching target
-    >>> zip(s_from, s_to[idx])
-    [(0, 0), (7, 10), (14, 10), (21, 20), (28, 30), (35, 30),
-     (42, 40), (49, 50), (56, 60), (63, 60), (70, 70), (77, 80),
-     (84, 80), (91, 90), (98, 90)]
-
-    Parameters
-    ----------
-    events_from : ndarray [shape=(n,)]
-      Array of events (eg, times, sample or frame indices) to match from.
-
-    events_to : ndarray [shape=(m,)]
-      Array of events (eg, times, sample or frame indices) to
-      match against.
-
-    left : bool
-    right : bool
-        If `False`, then matched events cannot be to the left (or right)
-        of source events.
-
-    Returns
-    -------
-    event_mapping : np.ndarray [shape=(n,)]
-        For each event in `events_from`, the corresponding event
-        index in `events_to`.
-
-        `event_mapping[i] == arg min |events_from[i] - events_to[:]|`
-
-    See Also
-    --------
-    match_intervals
-
-    Raises
-    ------
-    ParameterError
-        If either array of input events is not the correct shape
-    '''
-    if len(events_from) == 0 or len(events_to) == 0:
-        raise ParameterError('Attempting to match empty event list')
-
-    # If we can't match left or right, then only strict equivalence
-    # counts as a match.
-    if not (left or right) and not np.all(np.in1d(events_from, events_to)):
-            raise ParameterError('Cannot match events with left=right=False '
-                                 'and events_from is not contained '
-                                 'in events_to')
-
-    # If we can't match to the left, then there should be at least one
-    # target event greater-equal to every source event
-    if (not left) and max(events_to) < max(events_from):
-        raise ParameterError('Cannot match events with left=False '
-                             'and max(events_to) < max(events_from)')
-
-    # If we can't match to the right, then there should be at least one
-    # target event less-equal to every source event
-    if (not right) and min(events_to) > min(events_from):
-        raise ParameterError('Cannot match events with right=False '
-                             'and min(events_to) > min(events_from)')
-
-    # array of matched items
-    output = np.empty_like(events_from, dtype=np.int)
-
-    return __match_events_helper(output, events_from, events_to, left, right)
-
-
-@jit(nopython=True)
-def __match_events_helper(output, events_from, events_to, left=True, right=True):
-    # mock dictionary for events
-    from_idx = np.argsort(events_from)
-    sorted_from = events_from[from_idx]
-
-    to_idx = np.argsort(events_to)
-    sorted_to = events_to[to_idx]
-
-    # find the matching indices
-    matching_indices = np.searchsorted(sorted_to, sorted_from)
-
-    # iterate over indices in matching_indices
-    for ind, middle_ind in enumerate(matching_indices):
-        left_flag = False
-        right_flag = False
-
-        left_ind = -1
-        right_ind = len(matching_indices)
-
-        left_diff = 0
-        right_diff = 0
-        mid_diff = 0
-
-        middle_ind = matching_indices[ind]
-        sorted_from_num = sorted_from[ind]
-
-        # Prevent oob from chosen index
-        if middle_ind == len(sorted_to):
-            middle_ind -= 1
-
-        # Permitted to look to the left
-        if left and middle_ind > 0:
-            left_ind = middle_ind - 1
-            left_flag = True
-
-        # Permitted to look to right
-        if right and middle_ind < len(sorted_to) - 1:
-            right_ind = middle_ind + 1
-            right_flag = True
-
-        mid_diff = abs(sorted_to[middle_ind] - sorted_from_num)
-        if left and left_flag:
-            left_diff = abs(sorted_to[left_ind] - sorted_from_num)
-        if right and right_flag:
-            right_diff = abs(sorted_to[right_ind] - sorted_from_num)
-
-        if left_flag and (not right
-                and (sorted_to[middle_ind] > sorted_from_num) or
-                 (not right_flag and left_diff < mid_diff) or
-                 (left_diff < right_diff and left_diff < mid_diff)):
-            output[ind] = to_idx[left_ind]
-
-        # Check if right should be chosen
-        elif right_flag and (right_diff < mid_diff):
-            output[ind] = to_idx[right_ind]
-
-        # Selected index wins
-        else:
-            output[ind] = to_idx[middle_ind]
-
-    # Undo sorting
-    solutions = np.empty_like(output)
-    solutions[from_idx] = output
-
-    return solutions
 
 
 def localmax(x, axis=0):
